backend/pyproject.toml
[project]
name = "cape-town-rag-chat"
version = "0.1.0"
description = "Mini RAG chatbot about Cape Town and the Western Cape"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
